+++ test/quantum/zx.py
@@ -57,6 +57,7 @@ def test_Sum():
 
 def test_scalar():
     assert scalar(1j)[::-1] == scalar(-1j)
+    assert repr(scalar(0.5)) == "quantum.zx.Scalar(0.5)"
 
 
 def test_Functor():
@@ -189,3 +190,17 @@ def test_circuit2zx():
                 boxes_and_offsets=zip(
                     [SWAP, Z(1, 2), X(2, 1), scalar(2 ** 0.5), SWAP],
                     [1, 0, 1, 2, 1])))
+
+
+def test_strategy():
+    from discopy import testing
+    from discopy.quantum.zx import Diagram, Hadamard, Spider
+
+    testing.assert_strategy_finds(Diagram, Spider, Hadamard)
+
+
+def test_axioms():
+    from discopy import testing
+    from discopy.quantum.zx import Diagram
+
+    testing.assert_axioms(Diagram)